% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportData.R
\name{exportData}
\alias{exportData}
\title{exportData: Export data package views to .csv}
\usage{
exportData(filepath = NA, zip = FALSE)
}
\arguments{
\item{filepath}{Quoted path to save files to. If not specified, will save to your working directory.}

\item{zip}{Logical. If TRUE, exports a zip file. If FALSE (Default), exports individual csvs.}
}
\value{
NETN rocky intertidal views exported to specified path
}
\description{
This function exports data package views generated from the rocky intertidal front end
database individual .csv files, or as a zip file with the dataset name (NETN_RockyIntertidal_Data_Package),
and the date it was zipped. The exported .csv or .zip files can then be imported via importData().
This function is primarily for internal use to create flat files that users can import to run functions
in this package without having a connection having MS Access installed. Must first import the views in
the current R session.
}
\examples{
\dontrun{
# RUN FIRST
library(rockyIntertidal)
importData()

# Export csvs to working directory
exportData()

# Export a zip to the path specified
exportData(filepath = "../data/rocky", zip = TRUE)

# Export views as .csvs to specified path
exportData(filepath = "../data/rocky")
}

}
