% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterBolts.R
\name{filterBolts}
\alias{filterBolts}
\title{filterBolts: filter bolt data by park, location, plot type, or plot name}
\usage{
filterBolts(
  park = c("all"),
  location = c("all"),
  plotType = c("all"),
  species = c("all_records"),
  plotName = c("all")
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{location}{Include data from all locations, or choose specific locations based on location code.
\describe{
\item{'all'}{Includes all locations returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GRISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster}
}}

\item{plotType}{Type of plot to filter. Options include:
c("all", "Band Transect", "Benchmark", "Photoplot", "Point Intercept Transect",
"Recruitment Plot", "Reference", "Temperature Logger")}

\item{species}{Filter on target species. Note that "all_records" returns all options, whereas
"All" is part of the Target_Species lookup, and will return records where Target_Species == "All".
Options include:
c("all_records", "All", "Ascophyllum", "Barnacle", "Echinoderms", "Fucus", "Mussel",
"Red Algae", "Not Applicable")}

\item{plotName}{Filter on plot name. Options include:
c("A1", "A2", "A3", "A4", "A5", "B1", "B2", "B3", "B4", "B5", "BM",
"F1", "F2", "F3", "F4", "F5", "M1", "M2", "M3", "M4", "M5",
"R1", "R2", "R3", "R4", "R5", "REF", "S1", "S2", "S3", "S4", "S5",
"T1", "T2", "T3",
"TEMP1", "TEMP1 ", "TEMP2", "TEMP3",
"U1", "U2", "U3", "U4", "U5", "X1", "X2", "X3")}
}
\value{
Returns a data frame of bolt data filtered by function arguments
}
\description{
This function filter bolt data by park, location, plot type or plot name.
}
\examples{
\dontrun{

# Default filter returns all records
bolts <- filterBolts()

# Return photo plot bolts for all sites in BOHA
boha_photo <- filterBolts(park = "BOHA", plotType = "Photoplot")

# Return bolts for algal only in ACAD's Ship Harbor and Bass Harbor
acad_alg <- filterBolts(location = c("BASHAR", "SHIHAR"),
  species = c("Ascophyllum", "Fucus", "Red Algae"))

# Return bolts for barnacle recruitment plots in ACAD
acad_barn <- filterBolts(park = "ACAD", )

# Return bolts for summer barnacle recruitment photoplots at Little Hunter
lihu_sb <- filterBolts(location = "LITHUN", plotName = c("S1", "S2", "S3", "S4", "S5"))


# Return bolts for point intercept transects in Outer Brewster
ob_tr <- filterBolts(location = "OUTBRE", plotType = "Point Intercept Transect")

}


}
