% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBarnacleRecruitment.R
\name{getBarnacleRecruitment}
\alias{getBarnacleRecruitment}
\title{getBarnacleRecruitment: get Barnacle recruitment count data}
\usage{
getBarnacleRecruitment(
  park = "all",
  site = "all",
  plotName = "all",
  QAQC = FALSE,
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  dropNA = TRUE,
  timeTaken = "all"
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{plotName}{Filter on plot name. Options include:
c("all", "summer", "winter", "S1", "S2", "S3", "S4", "S5", "U1", "U2", "U3", "U4", "U5")}

\item{QAQC}{Logical. If FALSE (Default) does not return QAQC records. If TRUE,
returns all records, including QAQC scoring records. This differs from other functions in that
QAQC is determined at the record level, not the visit level.}

\item{years}{Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}

\item{dropNA}{Logical. If TRUE (default), blank counts are removed.
If FALSE, all records are returned.}

\item{timeTaken}{Filter on whether barnacle counts are from early (summer) or late (winter) photos from S-labeled plots.
Note that this filter is only valid for years 2019 and later. Options are "all" (default), "early", and "late". If early
or late is specified, records prior to 2019 will not be included in the results.}
}
\value{
Returns a data frame of barnacle recruitment count data.
}
\description{
This function filters barnacle recruitment count data by park, site, and plot name.
}
\examples{
\dontrun{

importData()

# Default filter returns all records except QAQC records and blank counts
barn <- getBarnacleRecruitment()

# Barnacle counts for ACAD only sites
barn_acad <- getBarnacleRecruitment(park = "ACAD")

# Barnacle counts for specific sites, plots, and years
barn_summer <- getBarnacleRecruitment(park = "ACAD", plotName = "summer")
barn_BOHA <- getBarnacleRecruitment(site = c("CALISL", "GREISL"))
barn_5yr <- getBarnacleRecruitment(years = 2016:2024)
barn_first_last <- getBarnacleRecruitment(years = c(2013, 2024))
barn21_qaqc <- getBarnacleRecruitment(years = 2024, QAQC = TRUE)

# Barnacle counts for winter-only photos taken of S plots.
barn_late <- getBarnacleRecruitment(years = 2019:2024, timeTaken = 'late')

}


}
