% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBolts.R
\name{getBolts}
\alias{getBolts}
\title{getBolts: get bolt data}
\usage{
getBolts(
  park = "all",
  site = "all",
  plotType = "all",
  plotName = "all",
  community = "all_records",
  coords = c("all"),
  dropNA = TRUE
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{plotType}{Type of plot to filter. Options include:
c("all", "Band Transect", "Benchmark", "Photoplot", "Point Intercept Transect",
"Recruitment Plot", "Reference", "Temperature Logger")}

\item{plotName}{Filter on plot name. Options include:
c("A1", "A2", "A3", "A4", "A5", "B1", "B2", "B3", "B4", "B5", "BM",
"F1", "F2", "F3", "F4", "F5", "M1", "M2", "M3", "M4", "M5",
"R1", "R2", "R3", "R4", "R5", "REF", "S1", "S2", "S3", "S4", "S5",
"T1", "T2", "T3",
"TEMP1", "TEMP1 ", "TEMP2", "TEMP3",
"U1", "U2", "U3", "U4", "U5", "X1", "X2", "X3")}

\item{community}{Filter on target community type. Note that "all_records" returns all options, whereas
"All" is part of the CommunityType lookup, and will return records where CommunityType == "All".
Options include:
c("all_records", "All", "Ascophyllum", "Barnacle", "Echinoderms", "Fucus", "Mussel",
"Red Algae", "Not Applicable")}

\item{coords}{Specify which GPS coordinates to return, Lat/Long, UTM Zone 19N, or both. Options:
c("latlong", "utm", "all")}

\item{dropNA}{Logical. If TRUE (default), blank X/Y coordinates and elevations are removed.
If FALSE, all records are returned.}
}
\value{
Returns a data frame of bolt data filtered by function arguments
}
\description{
This function filters bolt data by park, site, plot type or plot name.
}
\examples{
\dontrun{

importData()

# Default, returns all records except blank X/Y, or elevations
bolts <- getBolts()

# Return photo plot bolts for all sites in BOHA
boha_photo <- getBolts(park = "BOHA", plotType = "Photoplot")

# Return bolts for algal only in ACAD's Ship Harbor and Bass Harbor
acad_alg <- getBolts(site = c("BASHAR", "SHIHAR"),
  community = c("Ascophyllum", "Fucus", "Red Algae"))

# Return bolts for barnacle recruitment plots in ACAD
acad_barn <- getBolts(park = "ACAD", community = "Barnacle")

# Return bolts for summer barnacle recruitment photoplots at Little Hunter
lihu_sb <- getBolts(site = "LITHUN", plotName = c("S1", "S2", "S3", "S4", "S5"))

# Return bolts for point intercept transects in Outer Brewster
ob_tr <- getBolts(site = "OUTBRE", plotType = "Point Intercept Transect")

}


}
