% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEchinoCounts.R
\name{getEchinoCounts}
\alias{getEchinoCounts}
\title{getEchinoCounts: get Echinoderm count data}
\usage{
getEchinoCounts(
  park = "all",
  location = "all",
  plotName = "all",
  species = "all",
  QAQC = FALSE,
  years = 2013:as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{location}{Include data from all locations, or choose specific locations based on location code.
\describe{
\item{'all'}{Includes all locations returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster}
}}

\item{plotName}{Filter on plot name. Options include: c("all", "X1", "X2", and "X3")}

\item{species}{Filter on species code. Options include:
c("all", "ASTFOR", "ASTRUB", "HENSAN", "STRDRO"). If a new species is added,
the function will warn the user that an unrecognized species was specified in case it was an error.}

\item{QAQC}{Logical. If FALSE (Default) does not return QAQC events. If TRUE,
returns all events, including QAQC events.}

\item{years}{Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}
}
\value{
Returns a data frame of echinoderm count data in wide form filtered by function arguments.
}
\description{
This function filters echinoderm count data by park, location, and plot name. The returned
data frame is long, rather than wide, which is how it's imported. If a given species was not detected
during a visit, its count is 0.
Species codes are: ASTFOR = Asterias forbesi (Northern sea star); ASTRUB = Asterias rubens (common sea star);
HENSAN = Henricia sanguinolenta (blood sea star); STRDRO = Strongylocentrotus droebachiensis (sea urchin)
}
\examples{
\dontrun{

importData()

# Default filter returns all records
ech <- getEchinoCounts()

# Echino counts for ACAD only sites
ech_acad <- getEchinoCounts(park = "ACAD")

# Echino counts for specific sites, plots, species, and years

ech_t3 <- getEchinoCounts(park = "ACAD", plotName = "X3")
ech_BOHA2 <- getEchinoCounts(location = c("CALISL", "GREISL"))
ech_5yr <- getEchinoCounts(years = 2016:2021)
ech_first_last <- getEchinoCounts(years = c(2013, 2021))
ech21_qaqc <- getEchinoCounts(years = 2021, QAQC = TRUE)

}


}
