% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEchinoMeas.R
\name{getEchinoMeas}
\alias{getEchinoMeas}
\title{getEchinoMeas: get Echinoderm measurement data}
\usage{
getEchinoMeas(
  park = "all",
  site = "all",
  plotName = "all",
  species = "all",
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  QAQC = FALSE,
  dropNA = TRUE
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{plotName}{Filter on plot name. Options include: c("X1", "X2", and "X3")}

\item{species}{Filter on species code. Options include:
c("all", "ASTFOR", "ASTRUB", "HENSAN", "STRDRO")}

\item{years}{Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}

\item{QAQC}{Logical. If FALSE (Default) does not return QAQC events. If TRUE,
returns all events, including QAQC events.}

\item{dropNA}{Logical. If TRUE (default), blank  measurements are removed.
If FALSE, all records are returned.}
}
\value{
Returns a data frame of echinoderm measurement data filtered by function arguments.
}
\description{
This function filters echinoderm measurement data by park, site, plot name, and species. The returned data frame has a row for each species measured per sampling event and plot.
}
\examples{
\dontrun{

importData()

# Default filter returns all records except QAQC visits and blank measurements
ech <- getEchinoMeas()

# Echino counts for ACAD only sites
ech_acad <- getEchinoMeas(park = "ACAD")

# Echino counts for specific sites, plots, species, and years

ech_t3 <- getEchinoMeas(park = "ACAD", plotName = "X3")
ech_BOHA2 <- getEchinoMeas(site = c("CALISL", "GREISL"))
ech_5yr <- getEchinoMeas(years = 2016:2021)
ech_first_last <- getEchinoMeas(years = c(2013, 2021))
ech21_qaqc <- getEchinoMeas(years = 2021, QAQC = TRUE)
ech21_ASTFOR <- getEchinoMeas(species = "ASTFOR")

}


}
