% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotileInvertCounts.R
\name{getMotileInvertCounts}
\alias{getMotileInvertCounts}
\title{getMotileInvertCounts: get Motile Invertebrate counts}
\usage{
getMotileInvertCounts(
  park = "all",
  location = "all",
  plotName = "all",
  species = "all",
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  QAQC = FALSE
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{location}{Include data from all locations, or choose specific locations based on location code.
\describe{
\item{'all'}{Includes all locations returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster}
}}

\item{plotName}{Filter on plot name. Options include:
c("all", A1", "A2", "A3", "A4", "A5", "B1", "B2", "B3", "B4", "B5",
"F1", "F2", "F3", "F4", "F5", "M1", "M2", "M3", "M4", "M5",
"R1", "R2", "R3", "R4", "R5")}

\item{species}{Filter on species code. Options include:
c("all", "CARMAE", "HEMISAN", "LITLIT", "LITOBT", "LITSAX", "NUCLAP", "TECTES"). If a new species is added,
the function will warn the user that an unrecognized species was specified in case it was an error.}

\item{years}{Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}

\item{QAQC}{Logical. If FALSE (Default) does not return QAQC events. If TRUE,
returns all events, including QAQC events.}
}
\value{
Returns a data frame of motile invertebrate count data filtered by function arguments.
}
\description{
This function filters motile invertebrate count data by park, location, and plot name. The returned data frame lists counts of each species detected for each sampling event and plot. Species codes in the data set are: CARMAE = Carcinus maenas (green crab); HEMISAN = Hemigrapsus sanguineus (Asian shore crab); LITLIT = Littorina littorea (common periwinkle); LITOBT = Littorina obtusata (smooth periwinkle); LITSAX = Littorina saxatilis (rough periwinkle); NUCLAP = Nucella lapillus (dogwhelk), TECTES = Tectura testudinalis (limpet).
}
\examples{
\dontrun{

importData()

# Default filter returns all records
minv <- getMotileInvertCounts()

# Motile Invert counts for ACAD only sites
minv_acad <- getMotileInvertCounts(park = "ACAD")

# Motile Invert counts for specific sites, plots, species, and years

minv_r <- getMotileInvertCounts(park = "ACAD", plotName = c("R1", "R2", "R3", "R4", "R5"))
minv_BOHA2 <- getMotileInvertCounts(location = c("CALISL", "GREISL"))
minv_lit <- getMotileInvertCounts(species = c("LITLIT", "LITOBT", "LITSAX"))
minv_5yr <- getMotileInvertCounts(years = 2016:2021)
minv_first_last <- getMotileInvertCounts(years = c(2013, 2021))
minv21_qaqc <- getMotileInvertCounts(years = 2021, QAQC = TRUE)

}


}
