% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotileInvertMeas.R
\name{getMotileInvertMeas}
\alias{getMotileInvertMeas}
\title{getMotileInvertMeas: get Motile Invertebrate measurement data}
\usage{
getMotileInvertMeas(
  park = "all",
  site = "all",
  plotName = "all",
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  QAQC = FALSE,
  species = "all",
  community = "all",
  dropNA = TRUE
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{plotName}{Filter on plot name. Options include:
c("all", A1", "A2", "A3", "A4", "A5", "B1", "B2", "B3", "B4", "B5",
"F1", "F2", "F3", "F4", "F5", "M1", "M2", "M3", "M4", "M5",
"R1", "R2", "R3", "R4", "R5")}

\item{years}{Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}

\item{QAQC}{Logical. If FALSE (Default) does not return QAQC events. If TRUE,
returns all events, including QAQC events.}

\item{species}{Filter on species code. Options include:
c("all", "CARMAE", "HEMISAN", "LITLIT", "LITOBT", "LITSAX", "NUCLAP", "TECTES"). If a new species is added,
the function will warn the user that an unrecognized species was specified in case it was an error.}

\item{community}{Filter on target community. Options include:
c("Ascophyllum", "Barnacle", "Fucus", "Mussel", "Red Algae")}

\item{dropNA}{Logical. If TRUE (default), blank measurements are removed.
If FALSE, all records are returned.}
}
\value{
Returns a data frame of motile invertebrate measurement data filtered by function arguments.
}
\description{
This function filters motile invertebrate measurement data by park, site, and plot name.
The returned data frame lists measurement data of each species detected for each sampling event and plot.
Species codes in the data set are: CARMAE = Carcinus maenas (green crab); HEMISAN = Hemigrapsus sanguineus (Asian shore crab);
LITLIT = Littorina littorea (common periwinkle); LITOBT = Littorina obtusata (smooth periwinkle);
LITSAX = Littorina saxatilis (rough periwinkle); NUCLAP = Nucella lapillus (dogwhelk); TECTES = Tectura testudinalis (limpet).
}
\examples{
\dontrun{

importData()

# Default filter returns all records except QAQC visits and blank measurements
minv <- getMotileInvertMeas()

# Motile Invert measurement data for ACAD only sites
minv_acad <- getMotileInvertMeas(park = "ACAD")

# Motile Invert measurement data for specific sites, plots, species, and years

minv_r <- getMotileInvertMeas(park = "ACAD", plotName = c("R1", "R2", "R3", "R4", "R5"))
minv_BOHA2 <- getMotileInvertMeas(site = c("CALISL", "GREISL"))
minv_lit <- getMotileInvertMeas(species = c("LITLIT", "LITOBT", "LITSAX"))
minv_5yr <- getMotileInvertMeas(years = 2016:2021)
minv_first_last <- getMotileInvertMeas(years = c(2013, 2021))
minv21_qaqc <- getMotileInvertMeas(years = 2021, QAQC = TRUE)

}


}
