% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPIBoltDistance.R
\name{getPIBoltDistance}
\alias{getPIBoltDistance}
\title{getPIBoltDistance: get point intercept bolt distance}
\usage{
getPIBoltDistance(
  park = "all",
  site = "all",
  plotName = "all",
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  QAQC = FALSE,
  dropNA = TRUE
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{plotName}{Filter on plot name. Options include: c("all", "T1", "T2", and "T3")}

\item{years}{Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}

\item{QAQC}{Logical. If FALSE (Default) does not return QAQC events. If TRUE,
returns all events, including QAQC events.}

\item{dropNA}{Logical. If TRUE (default), blank distances are removed. If FALSE, all records are returned.}
}
\value{
Returns a data frame of point intercept bolt data filtered by function arguments
}
\description{
This function filters point intercept bolt data by park, site, plot name, and species.
}
\examples{
\dontrun{

importData()

# Default filter returns all records except QAQC visits and blank distances
bolt <- getPIBoltDistance()

# PI Bolt distances for ACAD only sites
bolt <- getPIBoltDistance(park = "ACAD")

# PI Bolt distances for specific parks, sites, plots, and years

bolt_t3 <- getPIBoltDistance(park = "ACAD", plotName = "T3")
bolt_BOHA2 <- getPIBoltDistance(site = c("CALISL", "GREISL"))
bolt_5yr <- getPIBoltDistance(years = 2016:2021)
bolt_first_last <- getPIBoltDistance(years = c(2013, 2021))
bolt_with_qaqc <- getPIBoltDistance(years = 2021, QAQC = TRUE)

}


}
