% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPISppDetections.R
\name{getPISppDetections}
\alias{getPISppDetections}
\title{getPISppDetections: get point intercept species detection data}
\usage{
getPISppDetections(
  park = "all",
  location = "all",
  plotName = "all",
  species = "all",
  years = 2013:as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{location}{Include data from all locations, or choose specific locations based on location code.
\describe{
\item{'all'}{Includes all locations returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster}
}}

\item{plotName}{Filter on plot name. Options include: c("T1", "T2", and "T3")}

\item{species}{Filter on species code. Options include:
c("all", "ALGBRO",  "ALGGRE", "ALGRED", "ARTCOR", "ASCEPI", "ASCNOD", "BARSPP",
"BOLT", "CHOMAS", "CRUCOR", "FUCEPI", "FUCSPP", "KELP", "MUSSPP", "NONCOR",
"OTHINV", "OTHSUB", "PALPAL", "PORSPP", "ROCK", "ULVENT", "ULVINT", "ULVLAC",
"UNIDEN", "WATER")}

\item{years}{Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}
}
\value{
Returns a data frame of point intercept species detection data filtered by function arguments
}
\description{
This function filters point intercept species detection data by park,
location, plot name, and species.
}
\examples{
\dontrun{

importData()

# Default filter returns all records
spp <- getPISppDetections()

# Species detections for ACAD only sites
spp <- getPISppDetections(park = "ACAD")

# Species detections for specific sites, plots, species, and years

spp_t3 <- getPISppDetections(park = "ACAD", plotName = "T3")
spp_BOHA2 <- getPISppDetections(location = c("CALISL", "GREISL"))
spp_fuc <- getPISppDetections(park = "BOHA", species = c("FUCEPI", "FUCSPP"))
spp_5yr <- getPISppDetections(years = 2016:2021)
spp_first_last <- getPISppDetections(years = c(2013, 2021))

}


}
