% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSppConditions.R
\name{getSppConditions}
\alias{getSppConditions}
\title{: getSppConditions}
\usage{
getSppConditions(
  park = "all",
  site = "all",
  years = 2013:as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}

#' @param years Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}
}
\value{
Dataframe with species conditions
}
\description{
This function filters recorded species conditions by park, site, and year.
}
\examples{
\dontrun{

importData()

# get species condition observations for BOHA in 2024

boha24 <- getSppConditions(park = "BOHA", years = 2024)

# get observations for Bass Harbor all years

bass <- getSppConditions(site = "BASHAR")

}

}
