% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{importData: Import views directly from NETN rocky intertidal database}
\usage{
importData(type = "DSN", odbc = "rocky_FE", filepath = NA, new_env = TRUE)
}
\arguments{
\item{type}{Select how to import the data package.
\describe{
\item{"DSN"}{Default. DSN database. If odbc is not specified, will default to rocky_FE.
Using this argument requires that you have a User DSN named rocky_FE that points to the
database containing the data package queries. Note that the database that generates
the data package views must also be linked to the latest rocky intertidal backend.}
\item{"dbfile"}{A specified database containing the data package queries for each view. If selected,
must provide the database filepath in the filepath argument.}
\item{"csv"}{Imports the csv version of the data package views. If selected, must provide the
filepath for the csvs in the filepath argument. This option that does not require MS Access and
ODBC driver on your computer.}
\item{"zip"}{Imports the csv versions of the data package views, as a zipped file. If selected,
must provide the filepath and name of the zip file. This option that does not require MS Access and
ODBC driver on your computer.}
}}

\item{odbc}{DSN of the database when using type = DSN. If not specified will default to "rocky_FE",
which should represent the database that generates the data package views.}

\item{filepath}{Quoted filepath where data package database (if type = "dbfile") or the csvs
(if type = "csv" or type = "zip") live.}

\item{new_env}{Logical. Specifies which environment to store views in. If \code{TRUE}(Default), stores
views in ROCKY environment. If \code{FALSE}, stores views in global environment}
}
\value{
Rocky intertidal database views in specified environment
}
\description{
This function imports views in the NETN rocky intertidal data package either as csv files or
queries in the rocky intertidal front end Access database. Each view is added to a ROCKY environment in
your workspace, or to your global environment based on whether new_env = TRUE or FALSE.
}
\examples{
\dontrun{

# Import views using DSN
importData(type ='DSN', odbc = "rocky_FE") # this is the same as importData()

# Import views from specified database:
importData(type ='dbfile',
  filepath = '../data/rocky/NETN_RockyIntertidal_FE_DataPackage_20241211.accdb')

# Import views from folder with csvs
importData(type = 'csv', filepath = "../data/rocky/")

# Import views from zip file of csvs
importData(type = 'zip',
  filepath = "../data/rocky/NETN_RockyIntertidal_Data_Package_20260917.zip")


}

}
