% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarnacleRecruitment.R
\name{plotBarnacleRecruitment}
\alias{plotBarnacleRecruitment}
\title{plotBarnacleRecruitment}
\usage{
plotBarnacleRecruitment(
  park = "all",
  site = "all",
  plotName = "summer",
  QAQC = FALSE,
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  plot_title = TRUE
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{plotName}{Filter on plot name. Options include:
c("all", "summer", "winter", "S1", "S2", "S3", "S4", "S5", "U1", "U2", "U3", "U4", "U5").
By default, only "summer" plots are included in the plots.}

\item{QAQC}{Logical. If FALSE (default) does not return QAQC records. If TRUE,
returns all records, including QAQC scoring records. This differs from other functions in that
QAQC is determined at the record level, not the visit level.}

\item{years}{Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}

\item{plot_title}{Logical. If TRUE (default), will add site name as a title to the plot. If FALSE,
not plot title will be added. Only enabled when one site is specified, as facets will include the
site name in the facet strip.}
}
\value{
Returns a ggplot object of percent cover from photos filtered by function arguments
}
\description{
This function plots median barnacle recruitment counts across the five summer plots per site and year,
with error bars representing min and max counts among the five plots. If only 1 plot is specified, then the raw
counts will be plotted. If both summer and winter plots are specified, resulting plot will be color-coded by season.
If more than one site and year is specified, results will be faceted by site.
}
\examples{
\dontrun{

importData()

# plot barnacle counts for ACAD summer counts for all sites and all years.
plotBarnacleRecruitment(park = "ACAD")

# plot summer and winter counts for Bass Harbor
plotBarnacleRecruitment(site = "BASHAR", plotName = "all")

# same as above, but drop plot title
plotBarnacleRecruitment(site = "BASHAR", plotName = "all", plot_title = F)

# plot summer counts for all sites in ACAD in 2023
plotBarnacleRecruitment(park = "ACAD", years = 2023)

# plot summer and winter counts for all BOHA sites
plotBarnacleRecruitment(park = "BOHA", plotName = "all")
}


}
