% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEchinoMeas.R
\name{plotEchinoMeas}
\alias{plotEchinoMeas}
\title{plotEchinoMeas: plots a heatmap of echinoderm by species, site and size class}
\usage{
plotEchinoMeas(
  park = "all",
  site = "all",
  plotName = "all",
  species = "all",
  xlab = "Year",
  ylab = "Length (mm)",
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  nrow = 1,
  plot_title = NULL,
  QAQC = FALSE
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{plotName}{Filter on plot name (transect). Options include:
c("all", "X1", "X2", "X3")}

\item{species}{Filter on species code. Options include:
c("all", "ASTFOR", "ASTRUB", "HENSAN", "STRDRO"). If a new species is added,
the function will warn the user that an unrecognized species was specified in case it was an error.}

\item{xlab}{Quoted text label for x axis. If not specified, defaults to 'Year'}

\item{ylab}{Quoted text label for y axis. If not specified, defaults to 'Length (mm)'}

\item{years}{Filter on year of data collected. Default is 2013 to current year.}

\item{nrow}{Number of rows in the heatmap facet. Default is 1. Only used when 1 site is selected.}

\item{plot_title}{If specified, plots the title on the figure. If NULL, no plot title included.}

\item{QAQC}{Logical. If FALSE (Default), does not return QAQC events. If TRUE,
returns all events, including QAQC events.}
}
\value{
Returns a ggplot object of number of records by measurment by year per species
}
\description{
This function plots a heatmap of the distribution of invertebrate size classes in 1 mm increments
by year for each specified site, target species photoplot, and species.
}
\examples{
\dontrun{

importData()

plotEchinoMeas(site = "SHIHAR", years = 2013:2019)

plotMotileInvertMeas(site = "BASHAR", species = "NUCLAP")

plotMotileInvertMeas(site = "CALISL", species = "LITLIT")


}


}
