% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMotileInvertMeas.R
\name{plotMotileInvertMeas}
\alias{plotMotileInvertMeas}
\title{plotMotileInvertMeas: plots a heatmap of motile invertebrates by species, site and size class}
\usage{
plotMotileInvertMeas(
  park = "all",
  site = "all",
  plotName = "all",
  species = "all",
  community = "all",
  xlab = "Year",
  ylab = "Length (mm)",
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  nrow = 1,
  plot_title = NULL,
  QAQC = FALSE
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{plotName}{Filter on plot name (transect). Options include:
c("all", A1", "A2", "A3", "A4", "A5", "B1", "B2", "B3", "B4", "B5",
"F1", "F2", "F3", "F4", "F5", "M1", "M2", "M3", "M4", "M5",
"R1", "R2", "R3", "R4", "R5")}

\item{species}{Filter on species code. Options include:
c("all", "CARMAE", "HEMISAN", "LITLIT", "LITOBT", "LITSAX", "NUCLAP", "TECTES"). If a new species is added,
the function will warn the user that an unrecognized species was specified in case it was an error.}

\item{community}{Filter on target community. Options include:
c("Ascophyllum", "Barnacle", "Fucus", "Mussel", "Red Algae")}

\item{xlab}{Quoted text label for x axis. If not specified, defaults to 'Year'}

\item{ylab}{Quoted text label for y axis. If not specified, defaults to 'Length (mm)'}

\item{years}{Filter on year of data collected. Default is 2013 to current year.}

\item{nrow}{Number of rows in the heatmap facet. Default is 1. Only used when 1 site is selected.}

\item{plot_title}{If specified, plots the title on the figure. If NULL, no plot title included.}

\item{QAQC}{Logical. If FALSE (Default), does not return QAQC events. If TRUE,
returns all events, including QAQC events.}
}
\value{
Returns a ggplot object of number of records by measurment by year per species
}
\description{
This function plots a heatmap of the distribution of invertebrate size classes in 1 mm increments
by year for each specified site, target species photoplot, and species. To ensure facets are working properly,
can either specify multiple sites or specify multiple communities, but not both. Not every combination of facet
may be plotted correctly. If the returned plot has multiple numbers on top of each other and mostly white cells,
try a more refined call, like only specifying 1 site and 1 community.
}
\examples{
\dontrun{

importData()

plotMotileInvertMeas(site = "BASHAR")

plotMotileInvertMeas(site = "SHIHAR", community = "Ascophyllum")

plotMotileInvertMeas(park = "BOHA", community = "Barnacle", plot_title = "BOHA Barnacle Plots")

plotMotileInvertMeas(site = "CALISL", species = "LITLIT", plot_title = "Littorina littorea (common periwinkle)")


}


}
