% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhotoCover.R
\name{plotPhotoCover}
\alias{plotPhotoCover}
\title{plotPhotoCover: plots average cover by species and year}
\usage{
plotPhotoCover(
  park = "all",
  site = "all",
  plotName = "all",
  species = "all",
  community = "all",
  heatmap = FALSE,
  top_spp = NULL,
  palette = c("default"),
  xlab = "Year",
  ylab = "\% Cover",
  main_groups = FALSE,
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  nrow = 1,
  plot_title = NULL,
  QAQC = FALSE,
  drop_missing = TRUE
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{plotName}{Filter on plot name (transect). Options include: c("all", "T1", "T2", and "T3")}

\item{species}{Filter on species code. Options include:
c("all", "ALGBRO",  "ALGGRE", "ALGRED", "ARTCOR", "ASCEPI", "ASCNOD", "BARSPP",
"CHOMAS", "CRUCOR", "FUCEPI", "FUCSPP", "KELP", "MUSSPP", "NONCOR", "NOSAMP",
"OTHINV", "OTHPLA", "OTHSUB", "PALPAL", "PORSPP", "ROCK", "SAND", "TAR",
"ULVINT", "ULVLAC", "UNIDEN"). If a new species is added, the function will warn the user
that an unrecognized species was specified in case it was an error.
If a new species is added, the function will warn the user that an unrecognized species was specified
in case it was an error. The default color palette will also need to be updated, if a new species is added.
The viridis palette will work with new species without adaptation, but labels will only be species codes.}

\item{community}{Filter on target species (ie photoplot). Options include:
c("Ascophyllum", "Barnacle", "Fucus", "Mussel", "Red Algae")}

\item{heatmap}{Logical. If FALSE (Default), will plot species cover (y) by year (x) as points with error bars,
and faceted on target species. If TRUE, will plot results as heat map with species as y, year by x, and color
values representing percent cover.}

\item{top_spp}{Integer. If specified, only plots the top n species by cover for each target species. If not
specified, plots all species, or only the species specified in the species argument. Note that species ties are
included, so could include more than the number specified. Species with less than 0.1\% total cover are
dropped when a top_spp is specified. This argument overrides the species argument.}

\item{palette}{Choices are "default" or "viridis". Default assigns logical colors to common species.
Viridis uses a color-blind friendly palette of blues, purples and yellows.}

\item{xlab}{Quoted text label for x axis. If not specified, defaults to 'Year'}

\item{ylab}{Quoted text label for y axis. If not specified, defaults to '\% Cover'}

\item{main_groups}{Logical. If TRUE, only plots red algae (combined Irish moss and red algae group), Fucus spp.,
Ascophyllum nodosum, mussels, and barnacles. If FALSE (Default), plots all species or only species specified. If species
are specified, this argument will override it and only plot main species groups.}

\item{years}{Filter on year of data collected. Default is 2013 to current year.}

\item{nrow}{Number of rows in the heatmap facet. Default is 1. Only used when 1 site is selected and heatmap = T.}

\item{plot_title}{Character. If specified, adds the specified text as a title on the figure. If NULL, no plot title included.}

\item{QAQC}{Logical. If FALSE (Default), does not return QAQC events. If TRUE,
returns all events, including QAQC events.}
}
\value{
Returns a ggplot object of percent cover from photos filtered by function arguments
}
\description{
This function plots median percent cover by species for a given park, site, years and
target species. The point for each species is the median cover across the photoplots that site, year
and target species. The error bars are the minimum and maximum cover recorded in photoplots for a target
species. Note that if more than 1 site is specified, more than one target species is specified, or both,
the resulting figure will facet on those variables. If elev = TRUE, data will be plotted by elevation
instead of plot name.
}
\examples{
\dontrun{

importData()

# Default filter returns a plot faceted on site and target species group
plotPhotoCover()

# Other variations
spp = c("ALGRED", "ASCNOD", "BARSPP", "NONCOR", "FUCSPP", "ULVLAC")
plotPhotoCover(site = "CALISL", palette = "default", species = spp)

# Return only the top 3 species for barnacle photo plots
plotPhotoCover(site = "SHIHAR", palette = "default", community = "Barnacle", top_spp = 3)

# Only plot main 5 species groups for BOHA sites
plotPhotoCover(park = "BOHA", main_groups = TRUE)

# Plot heatmap for main species groups for Green Island
plotPhotoCover(site = "GREISL", main_groups = TRUE, heatmap = TRUE)

# Plot top 4 species in barnacle photoplots only
plotPhotoCover(park = "ACAD", top_spp = 4, community = "Barnacle")

}


}
