% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPITransects.R
\name{plotPITransects}
\alias{plotPITransects}
\title{plotPITransects: plots transect distance by elevation by location, year and transect}
\usage{
plotPITransects(
  park = "all",
  location = "all",
  plotName = "all",
  xlab = "Distance (m)",
  ylab = "Elevation MLLW (m)",
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  QAQC = FALSE,
  drop_missing = TRUE
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{location}{Include data from all locations, or choose specific locations based on location code.
\describe{
\item{'all'}{Includes all locations returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster}
}}

\item{plotName}{Filter on plot name. Options include: c("all", "T1", "T2", and "T3")}

\item{xlab}{Quoted text label for x axis. If not specified, defaults to 'Distance (m)'}

\item{ylab}{Quoted text label for y axis. If not specified, defaults to 'Elevation MLLW (m)'}

\item{years}{Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}

\item{QAQC}{Logical. If FALSE (Default), does not return QAQC events. If TRUE,
returns all events, including QAQC events.}

\item{drop_missing}{Logical. If TRUE (Default), drops bolts with missing
elevation and distances and drops bolts that are not bolt 1 but have a distance of 0.}
}
\value{
Returns a data frame of point intercept species detection data filtered by function arguments
}
\description{
This function relates bolt elevation data with point intercept species detection data by park,
location, plot name, and species.
}
\examples{
\dontrun{

importData()

# Default filter returns all records
trans <- sumPItransDetections()

plotPITransects(park = "ACAD", location = "BASHAR", years = 2019)

}


}
