% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaterTemp.R
\name{plotWaterTemp}
\alias{plotWaterTemp}
\title{plotWaterTemp: plots water temperature at high tide time series}
\usage{
plotWaterTemp(
  park = "all",
  site = "all",
  palette = c("default"),
  xlab = "Year",
  ylab = "High Tide Water Temp (F)",
  gam = FALSE,
  facet = TRUE,
  plot_tmin = FALSE,
  plot_tmax = FALSE,
  years = 2011:as.numeric(format(Sys.Date(), "\%Y")),
  plot_title = NULL,
  facet_col = 1,
  legend_position = "bottom"
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Specify site to generate water temperature plot. If more than one site chosen, will
plot lines on the same plot unless facet = TRUE. Note that plotting data from all sites will be slow.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{palette}{Choices are "default", "viridis", "greyscale", or "black". Default assigns logical colors to common species.
Viridis uses a color-blind friendly palette of blues, purples and yellows.}

\item{xlab}{Quoted text label for x axis. If not specified, defaults to 'Year'}

\item{ylab}{Quoted text label for y axis. If not specified, defaults to 'High Tide Temp (F)'}

\item{gam}{Logical. If FALSE (default), only plots temperature values. If TRUE, plots a trend line
derived from generalize additive modelling. NOT CURRENTLY FUNCTIONAL}

\item{facet}{Logical. If TRUE, will plot sites in separate facets. FALSE (default) plots all sites
on one figure.}

\item{plot_tmin}{Logical. If TRUE, will plot a line connecting minimum recorded temperatures across years.}

\item{plot_tmax}{Logical. If TRUE, will plot a line connecting maximum recorded temperatures across years.}

\item{years}{Filter on year of data collected. Default is 2011 to current year.}

\item{plot_title}{If specified, plots the title on the figure. If NULL, no plot title included.}

\item{facet_col}{Numeric. Number of columns for the facet to plot. Defaults to 1.}

\item{legend_position}{Position of legend. Options are legend.position options in ggplot2. Default is 'bottom'.
For no legend, specify 'none'.}
}
\value{
Returns a ggplot object of water temperature data
}
\description{
This function plots water temperature at high tide by site. Note that function works on
compiled logger data, rather than working off the raw logger data. Compiled datasets should be imported
using importWaterTemp(). To speed up plotting of multiple sites or long time-series, plot the simplified
water temperature data derived from importWaterTemp(simplify = TRUE).
}
\examples{
\dontrun{

path <-
  "./Rocky_Intertidal/NETN/5_Data/Data_Files/Temperature/Compiled_HT_water_temps_2011-2022/"
importWaterTemp(path, simplify = TRUE, buoy = TRUE) # import water temp and buoy data and simplify to daily stats.

# Default filter returns a plot for BASHAR
plotWaterTemp()

# Other variations
plotWaterTemp(site = "CALISL", years = 2016:2022, plot_title = "Calf Island",
plot_tmax = T, plot_tmin = T)

plotWaterTemp(site = "SCHPOI", years = 2013:2022, palette = 'black')

}

}
