% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumBarnacleRecruitment.R
\name{sumBarnacleRecruitment}
\alias{sumBarnacleRecruitment}
\title{sumBarnacleRecruitment: sum Barnacle recruitment count data}
\usage{
sumBarnacleRecruitment(
  park = "all",
  site = "all",
  plotName = "all",
  QAQC = FALSE,
  years = 2013:as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}ed+56
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{plotName}{Filter on plot name. Options include:
c("all", "summer", "winter", "S1", "S2", "S3", "S4", "S5", "U1", "U2", "U3", "U4", "U5")}

\item{QAQC}{Logical. If FALSE (Default) does not return QAQC records. If TRUE,
returns all records, including QAQC scoring records. This differs from other functions in that
QAQC is determined at the record level, not the visit level.}

\item{years}{Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}
}
\value{
Returns a data frame of barnacle recruitment count data.
}
\description{
This function filters barnacle recruitment count data by park, site, and plot name and summarizes counts for summer and winter settlement periods.
}
\examples{
\dontrun{

importData()

# Default filter returns all records
barn <- sumBarnacleRecruitment()

# sum barnacle counts for ACAD only sites
barn_acad <- sumBarnacleRecruitment(park = "ACAD")

# sum barnacle counts for specific sites, plots,and years

barn_summer <- sumBarnacleRecruitment(park = "ACAD", plotName = "summer")
barn_BOHA <- sumBarnacleRecruitment(site = c("CALISL", "GREISL"))
barn_5yr <- sumBarnacleRecruitment(years = 2016:2021)
barn_first_last <- sumBarnacleRecruitment(years = c(2013, 2021))
barn21_qaqc <- sumBarnacleRecruitment(years = 2024, QAQC = TRUE)

}


}
