% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumEchinoCounts.R
\name{sumEchinoCounts}
\alias{sumEchinoCounts}
\title{sumEchinoCounts: summarize counts of echinoderms}
\usage{
sumEchinoCounts(
  park = "all",
  site = "all",
  plotName = "all",
  species = "all",
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  QAQC = FALSE
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{plotName}{Filter on plot name. Options include: c("all", "X1", "X2", and "X3")}

\item{species}{Filter on species code. Options include:
c("all", "ASTFOR", "ASTRUB", "HENSAN", "STRDRO"). If a new species is added, the
function will warn the user that an unrecognized species was specified in case it was an error.}

\item{years}{Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}

\item{QAQC}{Logical. If FALSE (Default) does not return QAQC events. If TRUE,
returns all events, including QAQC events. If TRUE, also returns QAQC_SameGrid and QAQC_NewGrid fields.}
}
\value{
Returns a data frame of site-level echinoderm data filtered by function arguments
}
\description{
This function summarizes site-level average, median, min and max counts
by park, site, plot name, and species.
}
\examples{
\dontrun{

importData()

# Default filter returns all records
cnts <- sumEchinoCounts ()

# Echino counts for ACAD only sites
cnts <- sumEchinoCounts (park = "ACAD")

# Echino counts for specific sites, plots, species, and years

cnt_a1 <- sumEchinoCounts (park = "ACAD", plotName = "X1")
cnt_BOHA2 <- sumEchinoCounts (site = c("CALISL", "GREISL"))
cnt_gc <- sumEchinoCounts (park = "BOHA", species = c("CARMAE"))
cnt_5yr <- sumEchinoCounts (years = 2016:2021)
cnt_first_last <- sumEchinoCounts (years = c(2013, 2021))
cnt21_with_qaqc <- sumEchinoCounts (years = 2021, QAQC = TRUE)

}


}
