% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumMotileInvertCounts.R
\name{sumMotileInvertCounts}
\alias{sumMotileInvertCounts}
\title{sumMotileInvertCounts: summarize counts of motile invertebrates}
\usage{
sumMotileInvertCounts(
  park = "all",
  site = "all",
  plotName = "all",
  species = "all",
  community = "all",
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  QAQC = FALSE
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{plotName}{Filter on plot name. Options include:
c("all", A1", "A2", "A3", "A4", "A5", "B1", "B2", "B3", "B4", "B5",
"F1", "F2", "F3", "F4", "F5", "M1", "M2", "M3", "M4", "M5",
"R1", "R2", "R3", "R4", "R5")}

\item{species}{Filter on species code. Options include:
c("all", "CARMAE", "HEMISAN", "LITLIT",  "LITOBT", "LITSAX", "NUCLAP", "TECTES").
If a new species is added, the function will warn the user
that an unrecognized species was specified in case it was an error.}

\item{community}{Filter on target community. Options include:
c("Ascophyllum", "Barnacle", "Fucus", "Mussel", "Red Algae")}

\item{years}{Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}

\item{QAQC}{Logical. If FALSE (Default) does not return QAQC events. If TRUE,
returns all events, including QAQC events. If TRUE, also returns QAQC_SameGrid and QAQC_NewGrid fields.}
}
\value{
Returns a data frame of site-level photoplot count data filtered by function arguments
}
\description{
This function summarizes site-level average, median, min and max counts
by park, site, plot name, and species.
}
\examples{
\dontrun{

importData()

# Default filter returns all records
cnts <- sumMotileInvertCounts ()

# Photoplot cover for ACAD only sites
cnts <- sumMotileInvertCounts (park = "ACAD")

# Species detections for specific sites, plots, species, and years

cnt_a1 <- sumMotileInvertCounts (park = "ACAD", plotName = "A1")
cnt_BOHA2 <- sumMotileInvertCounts (site = c("CALISL", "GREISL"))
cnt_gc <- sumMotileInvertCounts (park = "BOHA", species = c("CARMAE"))
cnt_5yr <- sumMotileInvertCounts (years = 2016:2021)
cnt_first_last <- sumMotileInvertCounts (years = c(2013, 2021))
cnt21_with_qaqc <- sumMotileInvertCounts (years = 2021, QAQC = TRUE)

}


}
