% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumPISpecies.R
\name{sumPISpecies}
\alias{sumPISpecies}
\title{sumPISpecies: summarize point intercept species detection data by species and elevation band}
\usage{
sumPISpecies(
  park = "all",
  site = "all",
  plotName = "all",
  species = "all",
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  QAQC = FALSE,
  drop_missing = TRUE
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{plotName}{Filter on plot name. Options include: c("all", "T1", "T2", and "T3")}

\item{species}{Filter on species code. Options include:
c("all", "ALGBRO",  "ALGGRE", "ALGRED", "ARTCOR", "ASCEPI", "ASCNOD", "BARSPP",
"BOLT", "CHOMAS", "CRUCOR", "FUCEPI", "FUCSPP", "KELP", "MUSSPP", "NONCOR",
"OTHINV", "OTHSUB", "PALPAL", "PORSPP", "ROCK", "ULVENT", "ULVINT", "ULVLAC",
"UNIDEN", "WATER"). If a new species is added, the function will warn the user
that an unrecognized species was specified in case it was an error.}

\item{years}{Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}

\item{QAQC}{Logical. If FALSE (Default), does not return QAQC events. If TRUE,
returns all events, including QAQC events.}

\item{drop_missing}{Logical. If TRUE (Default), drops bolts with missing
elevation and distances and drops bolts that are not bolt 1 but have a distance of 0.}
}
\value{
Returns a data frame of point intercept species detection data filtered by function arguments
}
\description{
This function relates bolt elevation data with point intercept species detection data by park,
site, plot name, and species. Returned values include elevation change between bolts, median elevation
between bolts, min/max elevation between bolts, distance between bolts, slope percent and degree between bolts,
horizontal distance between bolts, and interpolated elevation for point intercept sampling sites.
}
\examples{
\dontrun{

importData()

# Default filter returns all records
spp <- sumPISpecies()

# Species detections for ACAD only sites
spp <- sumPISpecies(park = "ACAD")

# Species detections for specific sites, plots, species, and years

spp_t3 <- sumPISpecies(park = "ACAD", plotName = "T3")
spp_BOHA2 <- sumPISpecies(site = c("CALISL", "GREISL"))
spp_fuc <- sumPISpecies(park = "BOHA", species = c("FUCEPI", "FUCSPP"))
spp_5yr <- sumPISpecies(years = 2016:2021)
spp_first_last <- sumPISpecies(years = c(2013, 2021))
spp21_with_qaqc <- sumPISpecies(years = 2021, QAQC = TRUE)
spp_no_drops <- sumPISpecies(drop_missing = FALSE)

}


}
