% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rocky.R
\name{theme_rocky}
\alias{theme_rocky}
\title{theme_rocky: custom ggplot2 theme for rockyIntertidal}
\usage{
theme_rocky()
}
\value{
This function must be used in conjunction with a ggplot object
}
\description{
This function customizes a theme for plotting data in the package, including removing the
default panel grids from ggplot2 figures.
}
\examples{
\dontrun{
importData()
spp_pi <- sumPISppDetections(drop_missing = T)
ggplot(spp_pi |> filter(Plot_Name == "T3"),
         aes(x = dist_bolt_first, y = elev_first,
             group = as.factor(Year), color = as.factor(Year))) +
       geom_line(lwd = 1) +
       facet_wrap(~SiteCode, scales = 'free') +
       scale_color_brewer("Year", type = 'div', palette = 'Dark2') +
       theme_rocky()
}
}
