% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBarnacleRecruitment.R
\name{getBarnacleRecruitment}
\alias{getBarnacleRecruitment}
\title{getBarnacleRecruitment: get Barnacle recruitment count data}
\usage{
getBarnacleRecruitment(
  park = "all",
  site = "all",
  plotName = "all",
  QAQC = FALSE,
  years = 2013:as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{plotName}{Filter on plot name. Options include:
c("all", "summer", "winter", "S1", "S2", "S3", "S4", "S5", "U1", "U2", "U3", "U4", "U5")}

\item{QAQC}{Logical. If FALSE (Default) does not return QAQC events. If TRUE,
returns all events, including QAQC events.}

\item{years}{Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}
}
\value{
Returns a data frame of barnacle recruitment count data.
}
\description{
This function filters barnacle recruitment count data by park, site, and plot name.
}
\examples{
\dontrun{

importData()

# Default filter returns all records
barn <- getBarnacleRecruitment()

# Barnacle counts for ACAD only sites
barn_acad <- getBarnacleRecruitment(park = "ACAD")

# Barnacle counts for specific sites, plots, species, and years

barn_summer <- getBarnacleRecruitment(park = "ACAD", plotName = "summer")
barn_BOHA <- getBarnacleRecruitment(site = c("CALISL", "GREISL"))
barn_5yr <- getBarnacleRecruitment(years = 2016:2021)
barn_first_last <- getBarnacleRecruitment(years = c(2013, 2021))
barn21_qaqc <- getBarnacleRecruitment(years = 2021, QAQC = TRUE)

}


}
