% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEvents.R
\name{getEvents}
\alias{getEvents}
\title{getEvents: get specified events}
\usage{
getEvents(
  park = "all",
  site = "all",
  plotName = "all",
  plotType = "all",
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  QAQC = FALSE
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{plotName}{Filter on plot name. Options include:
c("A1", "A2", "A3", "A4", "A5", "B1", "B2", "B3", "B4", "B5", "BM",
"F1", "F2", "F3", "F4", "F5", "M1", "M2", "M3", "M4", "M5",
"R1", "R2", "R3", "R4", "R5", "REF", "S1", "S2", "S3", "S4", "S5",
"T1", "T2", "T3",
"TEMP1", "TEMP1 ", "TEMP2", "TEMP3",
"U1", "U2", "U3", "U4", "U5", "X1", "X2", "X3")}

\item{plotType}{Type of plot to filter. Options include:
c("all", "Band Transect", "Benchmark", "Photoplot", "Point Intercept Transect",
"Recruitment Plot", "Reference", "Temperature Logger")}

\item{years}{Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}

\item{QAQC}{Logical. If FALSE (Default) does not return QAQC events. If TRUE,
returns all events, including QAQC events.}
}
\value{
Returns a data frame of point intercept species detection data filtered by function arguments
}
\description{
This function filters events and returns notes recorded during sampling events by park,
site, plot name, and species.
}
\examples{
\dontrun{

importData()

# Default filter returns all records
notes <- getEvents()

# Event Notes for ACAD only sites
notes_ACAD <- getEvents(park = "ACAD")

}


}
