% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{importData: Import views directly from NETN rocky intertidal database}
\usage{
importData(
  type = c("DSN", "file"),
  DSN = "rocky_BE",
  path = NA,
  new_env = TRUE
)
}
\arguments{
\item{type}{Specify whether you are connecting to a DSN or file.
\describe{
\item{"DSN"}{Default. DSN database. If odbc argument is not specified, will default to rocky_BE.}
\item{"file"}{Connects to a backend based on the specified path}}}

\item{DSN}{DSN name of the database when using type = DSN. If not specified will defaut to "rocky_BE".}

\item{path}{Quoted path of database backend file, including the name of the backend.}

\item{new_env}{Logical. Specifies which environment to store views in. If \code{TRUE}(Default), stores
views in ROCKY environment. If \code{FALSE}, stores views in global environment}
}
\value{
Rocky intertidal database views in specified environment
}
\description{
This function imports flat files from the rocky intertidal database. Each view
is added to a ROCKY environment in your workspace, or to your global environment based on whether
new_env = TRUE or FALSE. You must have the latest ODBC SQL driver installed for this function to
work. It can be downloaded from: https://go.microsoft.com/fwlink/?linkid=2168524
}
\examples{
\dontrun{

# Import views using DSN
importData(type='DSN', DSN="rocky_BE") # this is the same as importData()

# Import database in specific folder:
importData(type='file', path='./Data/NETN_RockyIntertidal_Database_be_20230120.mdb')

}

}
