% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importWaterTemp.R
\name{importWaterTemp}
\alias{importWaterTemp}
\title{importWaterTemp: Import compiled water temperature data}
\usage{
importWaterTemp(path = NA, simplify = TRUE, new_env = TRUE)
}
\arguments{
\item{path}{Quoted path of database backend file, including the name of the backend.}

\item{simplify}{Logical. If TRUE (default), will simplify dataset to only include 1 water temperature measurement per high
tide event by selecting the value closest in time to high tide. This will speed up processing and plotting time. If FALSE,
will keep all values that are within 2 hours of high tide.}

\item{new_env}{Logical. Specifies which environment to store views in. If \code{TRUE}(Default), stores
views in ROCKY environment. If \code{FALSE}, stores views in global environment}
}
\value{
Rocky intertidal compiled water temperature data in specified environment
}
\description{
This function imports compiled water temperature data collected within 2 hours of high tide.
User must specify the path where the compiled logger data are stored, then imports and defines date fields
as POSIXct dates.
}
\examples{
\dontrun{

# Import views using DSN
importData(type='DSN', DSN="rocky_BE") # this is the same as importData()

# Import database in specific folder:
importData(type='file', path='./Data/NETN_RockyIntertidal_Database_be_20230120.mdb')

}

}
