% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBuoyData.R
\name{plotBuoyData}
\alias{plotBuoyData}
\title{plotBuoyData: plots daily buoy statistics}
\usage{
plotBuoyData(
  park = "ACAD",
  palette = c("default"),
  metric = "all",
  years = 2011:as.numeric(format(Sys.Date(), "\%Y")),
  plot_title = NULL,
  facet_col = 1
)
}
\arguments{
\item{park}{Choose the park to plot. Can only plot one park at a time.
\describe{
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{palette}{Choices are "default", "viridis", or "black". Default assigns specific colors to each metric.
Viridis uses a color-blind friendly palette of blues, purples and yellows.}

\item{metric}{Select the metric to plot. Options are below. Can either choose 1 plot or all plots.
\describe{
\item{'all'}{Plot individual facet for each metric.}
\item{'temp'}{Minimum and Maximum daily water temperature in F}
\item{'wspd'}{Maximum daily windspeed in miles/hour and wind direction for winds > 40 miles/hour.}
}}

\item{years}{Filter on year of data collected. Default is 2011 to current year.}

\item{plot_title}{If specified, plots the title on the figure. If NULL, no plot title included.}

\item{facet_col}{Numeric. Number of columns for the facet to plot. Defaults to 1.}
}
\value{
Returns a ggplot object of water temperature data
}
\description{
This function plots water temperature at high tide by location. Note that function works on
compiled logger data, rather than working off the raw logger data. Compiled datasets should be imported
using importWaterTemp(). To speed up plotting of multiple locations or long time-series, plot the simplified
water temperature data derived from importWaterTemp(simplify = TRUE).
}
\examples{
\dontrun{

path <-
  "Z:/PROJECTS/MONITORING/Rocky_Intertidal/NETN/5_Data/Data_Files/Temperature/Compiled_HT_water_temps_2011-2022/"
importWaterTemp(path, simplify = TRUE, buoy = TRUE) # import water temp and buoy data and simplify to daily stats.

# Default filter returns a plot of the closest buoys to ACAD and BOHA
plotBuoyData()

# Other variations
# Windspeed and direction only
plotBuoyData(metric = 'wspd'))

plotWaterTemp(location = "SCHPOI", years = 2013:2022, palette = 'black')

}

}
