% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEchinoCounts.R
\name{plotEchinoCounts}
\alias{plotEchinoCounts}
\title{plotEchinoCounts: summary plot of counts by species by year}
\usage{
plotEchinoCounts(
  park = "all",
  location = "all",
  plotName = "all",
  species = "all",
  palette = c("default"),
  xlab = "Year",
  ylab = "Average Count",
  main_groups = FALSE,
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  nrow = 1,
  plotly = F,
  xaxis = TRUE,
  plot_title = NULL,
  QAQC = FALSE,
  drop_missing = TRUE
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{location}{Include data from all locations, or choose specific locations based on location code.
\describe{
\item{'all'}{Includes all locations returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster}
}}

\item{plotName}{Filter on plot name (transect). Options include: c("all", "X1", "X2", and "X3")}

\item{species}{Filter on species code. Options include:
c("all", "ASTFOR", "ASTRUB", "HENSAN", "STRDRO"). If a new species is added, the
function will warn the user that an unrecognized species was specified in case it was an error.}

\item{palette}{Choices are "default" or "viridis". Default assigns logical colors to common species.
Viridis uses a color-blind friendly palette of blues, purples and yellows.}

\item{xlab}{Quoted text label for x axis. If not specified, defaults to 'Year'}

\item{ylab}{Quoted text label for y axis. If not specified, defaults to '\% Cover'}

\item{years}{Filter on year of data collected. Default is 2013 to current year.}

\item{plotly}{Logical. If TRUE, converts ggplot object to plotly object and includes tooltips. If FALSE (default),
plots a ggplot object.}

\item{plot_title}{If specified, plots the title on the figure. If blank, no plot title included.}

\item{QAQC}{Logical. If FALSE (Default), does not return QAQC events. If TRUE,
returns all events, including QAQC events.}
}
\value{
Returns a ggplot object of percent cover from photos filtered by function arguments
}
\description{
This function plots average counts by species for a given park, location, years and
species. The point for each species is the median count across the tidepools for that site, year
and pecies. The error bars are the middle 50\% (lower 25\% and upper 75\% quantiles)
recorded in tidepools for a species. Note that if more than 1 location is specified,
the resulting figure will facet on those variables.
}
\examples{
\dontrun{

importData()

# Default filter returns a plot faceted on location and species
plotEchinoCounts()

# Other variations

plotEchinoCounts(park = "ACAD", location = "SHIHAR")

stars = c("ASTFOR", "ASTRUB", "HENSAN")
plotEchinoCounts(location = "CALISL", palette = "default", species = stars)

}


}
