% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMotileInvertCounts.R
\name{plotMotileInvertCounts}
\alias{plotMotileInvertCounts}
\title{plotMotileInvertCounts: summary plot of counts by species and year}
\usage{
plotMotileInvertCounts(
  park = "all",
  site = "all",
  plotName = "all",
  species = "all",
  category = "all",
  community = "all",
  top_spp = NULL,
  palette = c("default"),
  xlab = "Year",
  ylab = "Median Count",
  main_groups = FALSE,
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  nrow = 1,
  plotly = F,
  xaxis = TRUE,
  plot_title = NULL,
  QAQC = FALSE,
  drop_missing = TRUE
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{plotName}{Filter on plot name. Options include:
c("all", A1", "A2", "A3", "A4", "A5", "B1", "B2", "B3", "B4", "B5",
"F1", "F2", "F3", "F4", "F5", "M1", "M2", "M3", "M4", "M5",
"R1", "R2", "R3", "R4", "R5")}

\item{species}{Filter on species code. Options include:
c("all", "CARMAE", "HEMISAN", "LITLIT",  "LITOBT", "LITSAX", "NUCLAP", "TECTES").
If a new species is added, the function will warn the user
that an unrecognized species was specified in case it was an error.}

\item{community}{Filter on target community. Options include:
c("Ascophyllum", "Barnacle", "Fucus", "Mussel", "Red Algae")}

\item{palette}{Choices are "default" or "viridis". Default assigns logical colors to common species.
Viridis uses a color-blind friendly palette of blues, purples and yellows.}

\item{xlab}{Quoted text label for x axis. If not specified, defaults to 'Year'}

\item{ylab}{Quoted text label for y axis. If not specified, defaults to '\% Cover'}

\item{years}{Filter on year of data collected. Default is 2013 to current year.}

\item{plotly}{Logical. If TRUE, converts ggplot object to plotly object and includes tooltips. If FALSE (default),
plots a ggplot object.}

\item{xaxis}{Logical. If TRUE (Default), plots years on x axis. If FALSE, drops x axis.
FALSE is useful for combining plotly photocover plots in a subplot.}

\item{plot_title}{If specified, plots the title on the figure. If blank, no plot title included.}

\item{QAQC}{Logical. If FALSE (Default), does not return QAQC events. If TRUE,
returns all events, including QAQC events.}
}
\value{
Returns a ggplot object of counts from photoplots filtered by function arguments
}
\description{
This function plots median count by species for a given park, site, years and
target species. The point for each species is the median count across the photoplots that site, year
and target species. The error bars are the middle 50\% (lower 25\% and upper 75\% quantiles)
recorded in photoplots for a target species. Note that if more than 1 site is specified,
more than one target species is specified, or both, the resulting figure will facet on those
variables.
}
\examples{
\dontrun{

importData()

# Default filter returns a plot faceted on site and target species group
plotMotileInvertCounts()

# Other variations

plotMotileInvertCounts(park = "ACAD", community = "Barnacle")

spp = c("CARMAE", "HEMISAN")
plotMotileInvertCounts(site = "CALISL", palette = "default", title = FALSE,
               species = spp)

}


}
