% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhotoCover.R
\name{plotPhotoCover}
\alias{plotPhotoCover}
\title{plotPhotoCover: plots average cover by species and year}
\usage{
plotPhotoCover(
  park = "all",
  location = "all",
  plotName = "all",
  species = "all",
  category = "all",
  target_species = "all",
  top_spp = NULL,
  palette = c("default"),
  xlab = "Year",
  ylab = "\% Cover",
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  plot_title = NULL,
  QAQC = FALSE,
  drop_missing = TRUE
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{location}{Include data from all locations, or choose specific locations based on location code.
Can only specify one location for each function call.}

\item{plotName}{Filter on plot name (transect). Options include: c("all", "T1", "T2", and "T3")}

\item{species}{Filter on species code. Options include:
c("all", "ALGBRO",  "ALGGRE", "ALGRED", "ARTCOR", "ASCEPI", "ASCNOD", "BARSPP",
"CHOMAS", "CRUCOR", "FUCEPI", "FUCSPP", "KELP", "MUSSPP", "NONCOR", "NOSAMP",
"OTHINV", "OTHPLA", "OTHSUB", "PALPAL", "PORSPP", "ROCK", "SAND", "TAR",
"ULVINT", "ULVLAC", "UNIDEN"). If a new species is added, the function will warn the user
that an unrecognized species was specified in case it was an error.
If a new species is added, the function will warn the user that an unrecognized species was specified
in case it was an error. The default color palette will also need to be updated, if a new species is added.
The viridis palette will work with new species without adaptation, but labels will only be species codes.}

\item{category}{Filter on category. Options include:
c("all", "Genus", "Species", "Species Group", and "Substrate")}

\item{target_species}{Filter on target species (ie photoplot). Options include:
c("Ascophyllum", "Barnacle", "Fucus", "Mussel", "Red Algae")}

\item{top_spp}{Integer. If specified, only plots the top n species by cover for each target species. If not
specified, plots all species, or only the species specified in the species argument. Note that species ties are
included, so could include more than the number specified. Species with less than 0.1\% total cover are also
dropped when a top_spp is specififed.}

\item{palette}{Choices are "default" or "viridis". Default assigns logical colors to common species.
Viridis uses a color-blind friendly palette of blues, purples and yellows.}

\item{xlab}{Quoted text label for x axis. If not specified, defaults to 'Year'}

\item{ylab}{Quoted text label for y axis. If not specified, defaults to '\% Cover'}

\item{years}{Filter on year of data collected. Default is 2013 to current year.

\describe{
\item{'all'}{Includes all locations returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster}
}}

\item{plot_title}{If specified, plots the title on the figure. If NULL, no plot title included.}

\item{QAQC}{Logical. If FALSE (Default), does not return QAQC events. If TRUE,
returns all events, including QAQC events.}
}
\value{
Returns a ggplot object of percent cover from photos filtered by function arguments
}
\description{
This function plots median percent cover by species for a given park, location, years and
target species. The point for each species is the median cover across the photoplots that site, year
and target species. The error bars are the minimum and maximum cover recorded in photoplots for a target
species. Note that if more than 1 location is specified, more than one target species is specified, or both,
the resulting figure will facet on those variables.
}
\examples{
\dontrun{

importData()

# Default filter returns a plot for BASHAR (have to pick a site)
plotPhotoCover()

# Other variations
spp = c("ALGRED", "ASCNOD", "BARSPP", "NONCOR", "FUCSPP", "ULVLAC")
plotPhotoCover(location = "CALISL", palette = "default", title = FALSE, facet = TRUE,
               species = spp)

plotPhotoCover(location = "SHIHAR", palette = "default",
               facet = FALSE, species = c("BARSPP"))

}


}
