% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpeciesContours.R
\name{plotSpeciesContours}
\alias{plotSpeciesContours}
\title{plotSpeciesContours: plots point intercept species ranges and photocover as pie charts}
\usage{
plotSpeciesContours(
  site = "BASHAR",
  palette = "default",
  xlab = "Distance (m)",
  ylab = "Elevation MLLW (m)",
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  plot_title = TRUE,
  QAQC = FALSE,
  include_photoplots = TRUE
)
}
\arguments{
\item{site}{Must choose a site to plot
\describe{
\item{"BASHAR"}{Bass Harbor, ACAD (default)}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{palette}{Choices are "default" or "viridis". Default assigns logical colors to common species.
Viridis uses a color-blind friendly palette of blues, purples and yellows.}

\item{years}{Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}

\item{plot_title}{Logical. If TRUE (default), plots site code as plot title. If FALSE, doesn't include a title.}

\item{include_photoplots}{Logical. If TRUE (default), includes pie charts of photoplot cover by elevation and distance.
If FALSE, only plots species contours.}
}
\value{
Returns a ggplot object of point intercept and percent cover data filtered by function arguments
}
\description{
This function plots a loess smoothed contour averaging the transects
across all years specified. Point intercept minimum and maximum elevation ranges
are plotted along the contours by year for each of the main species groups:
(REDGRP, ASCNOD, FUCSPP, MUSSPP, BARSPP, NONCOR). Photoplot cover is plotted as
median cover and median elevation for each target species plot.
}
\examples{
\dontrun{

importData()

# Default filter returns a plot faceted on site and target species group
plotSpeciesContours(site = "OTTPOI")


# Drop photoplots from plot
plotSpeciesContours(site = "CALISL", include_photoplots = F)

# Other variations
plotSpeciesContours(site = "CALISL", years = 2019)

plotSpeciesContours(site = "BASHAR", years = c(2013, 2018, 2023))

}


}
