% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaterTemp.R
\name{plotWaterTemp}
\alias{plotWaterTemp}
\title{plotWaterTemp: plots water temperature at high tide time series}
\usage{
plotWaterTemp(
  park = "all",
  location = "all",
  palette = c("default"),
  xlab = "Year",
  ylab = "High Tide Water Temp (F)",
  gam = FALSE,
  facet = TRUE,
  years = 2011:as.numeric(format(Sys.Date(), "\%Y")),
  plot_title = NULL
)
}
\arguments{
\item{park}{Include data from all parks, or choose one.
\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}}

\item{location}{Specify location to generate water temperature plot. If more than one location chosen, will
plot lines on the same plot unless facet = TRUE. Note that plotting data from all locations will be slow.
\describe{
\item{'all'}{Includes all locations returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster}
}}

\item{palette}{Choices are "default", "viridis", or "black". Default assigns logical colors to common species.
Viridis uses a color-blind friendly palette of blues, purples and yellows.}

\item{xlab}{Quoted text label for x axis. If not specified, defaults to 'Year'}

\item{ylab}{Quoted text label for y axis. If not specified, defaults to 'High Tide Temp (F)'}

\item{gam}{Logical. If FALSE (default), only plots temperature values. If TRUE, plots a trend line
derived from generalize additive modelling. NOT CURRENTLY FUNCTIONAL}

\item{facet}{Logical. If TRUE, will plot locations in separate facets. FALSE (default) plots all locations
on one figure.}

\item{years}{Filter on year of data collected. Default is 2011 to current year.}

\item{plot_title}{If specified, plots the title on the figure. If NULL, no plot title included.}
}
\value{
Returns a ggplot object of water temperature data
}
\description{
This function plots water temperature at high tide by location. Note that function works on
compiled logger data, rather than working off the raw logger data. Compiled datasets should be imported
using importWaterTemp(). To speed up plotting of multiple locations or long time-series, plot the simplified
water temperature data derived from importWaterTemp(simplify = TRUE).
}
\examples{
\dontrun{

path <- "Z:/PROJECTS/MONITORING/Rocky_Intertidal/NETN/5_Data/Data_Files/Temperature/Compiled_HT_water_temps_2011-2022/"
importWaterTemp(path, simplify = TRUE)

# Default filter returns a plot for BASHAR
plotWaterTemp()

# Other variations
plotWaterTemp(location = "CALISL", years = 2016:2022, plot_title = "Calf Island")

# Plot gam trend line (not functional yet)
plotWaterTemp(location = "SHIHAR", years = 2011:2022, plot_title = "Ship Harbor", gam = T)

}

}
