% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumEchinoMeas.R
\name{sumEchinoMeas}
\alias{sumEchinoMeas}
\title{sumEchinoMeas: summarize Echinoderm measurement data by species and 5mm size clases}
\usage{
sumEchinoMeas(
  park = "all",
  site = "all",
  plotName = "all",
  years = 2013:as.numeric(format(Sys.Date(), "\%Y")),
  QAQC = FALSE,
  species = "all"
)
}
\arguments{
\item{site}{Include data from all sites, or choose specific sites based on site code.
\describe{
\item{'all'}{Includes all sites returned by other filter arguments in function}
\item{"BASHAR"}{Bass Harbor, ACAD}
\item{"LITHUN"}{Little Hunter, ACAD}
\item{"LITMOO"}{Little Moose, ACAD}
\item{"OTTPOI"}{Otter Point, ACAD}
\item{"SCHPOI"}{Schoodic Point, ACAD}
\item{"SHIHAR"}{Ship Harbor, ACAD}
\item{"CALISL"}{Calf Island, BOHA}
\item{"GREISL"}{Green Island, BOHA}
\item{"OUTBRE"}{Outer Brewster, BOHA}
}}

\item{plotName}{Filter on plot name. Options include:
c("all", "X1", "X2", "X3")}

\item{years}{Filter on year of data collected. Default is 2013 to current year.
Can specify a vector of years.}

\item{QAQC}{Logical. If FALSE (Default) does not return QAQC events. If TRUE,
returns all events, including QAQC events.}

\item{species}{Filter on species code. Options include:
c("all", "ASTFOR", "ASTRUB", "HENSAN", "STRDRO"). If a new species is added,
the function will warn the user that an unrecognized species was specified in case it was an error.}
}
\value{
Returns a data frame of summarized echinoderm measurement data filtered by function arguments.
}
\description{
This function summarizes number of individuals in each 5 mm size class by park, site, and species.

\describe{
\item{'all'}{Includes all parks monitored in the network}
\item{'ACAD'}{Includes only sites in Acadia National Park}
\item{'BOHA'}{Includes only sites in Boston Harbor Islands National Recreation Area}
}
}
\examples{
\dontrun{

importData()

# Default filter returns all records
minv <- sumEchinoMeas()

# Motile Invert measurement data for ACAD only sites
minv_acad <- sumEchinoMeas(park = "ACAD")

# Motile Invert measurement data for specific sites, plots, species, and years

minv_r <- sumEchinoMeas(park = "ACAD", plotName = c("X1", "X2"))
minv_BOHA2 <- sumEchinoMeas(site = c("CALISL", "GREISL"))
minv_lit <- sumEchinoMeas(species = c("ASTFOR", "ASTRUB"))
minv_5yr <- sumEchinoMeas(years = 2016:2021)
minv_first_last <- sumEchinoMeas(years = c(2013, 2021))
minv21_qaqc <- sumEchinoMeas(years = 2021, QAQC = TRUE)

}


}
